Authorization Basics

+++ app/Http/Controllers/IdeaController.php
@@ -12,6 +12,8 @@ class IdeaController extends Controller
       'content' => 'required|min:5|max:240',
     ]);
 
+    $validate['user_id'] = auth()->id();
+
     Idea::create($validate);
     return redirect()->route('dashboard')->with('success', 'Idea created successfully');
   }
@@ -23,12 +25,20 @@ class IdeaController extends Controller
 
   public function edit(Idea $idea)
   {
+    if (auth()->id() !== $idea->user_id) {
+      abort(404, 'Idea not found');
+    }
+    
     $editing = true;
     return view('idea.show', compact('idea', 'editing'));
   }
 
   public function update(Idea $idea)
   {
+    if (auth()->id() !== $idea->user_id) {
+      abort(404, 'Idea not found');
+    }
+
     $validate = request()->validate([
       'content' => 'required|min:5|max:240',
     ]);
@@ -39,6 +49,10 @@ class IdeaController extends Controller
 
   public function destroy(Idea $idea)
   {
+    if (auth()->id() !== $idea->user_id) {
+      abort(404, 'Idea not found');
+    }
+
     $idea->delete();
     return redirect()->route('dashboard')->with('success', 'Idea deleted successfully');
   }
